app.py
import streamlit as st
import os
import gdown
import torch
from PIL import Image

# -------------------------------
# 📌 DOWNLOAD MODEL (1GB)
# -------------------------------
MODEL_PATH = "best_model.pth"

if not os.path.exists(MODEL_PATH):
    st.write("Downloading model... please wait ⏳")
    url = "https://drive.google.com/uc?id=1LCfciCjUA77qF_gwL5OyOxES2AAo0HCM"
    gdown.download(url, MODEL_PATH, quiet=False)

# -------------------------------
# 📌 LOAD MODEL (CACHE)
# -------------------------------
@st.cache_resource
def load_model():
    model = torch.load(MODEL_PATH, map_location="cpu")
    model.eval()
    return model

model = load_model()

# -------------------------------
# 📌 UI
# -------------------------------
st.title("Cross-Lingual Visual Question Answering System")

# Upload image
uploaded_file = st.file_uploader("Upload an image", type=["jpg", "png"])

# Input question
question = st.text_input("Ask a question (English or Telugu)")

# -------------------------------
# 📌 PREDICTION
# -------------------------------
if uploaded_file is not None:
    image = Image.open(uploaded_file)
    st.image(image, caption="Uploaded Image", use_container_width=True)

    if question:
        try:
            # 🔥 TODO: Replace with your real model logic
            # Example:
            # image_tensor = preprocess_image(image)
            # question_tensor = preprocess_question(question)
            # output = model(image_tensor, question_tensor)
            # pred = torch.argmax(output, dim=1).item()
            # answer = idx2answer[pred]

            answer = "Demo Answer"  # temporary

        except Exception as e:
            st.error(f"Error: {e}")
            answer = "Prediction failed"

        st.subheader("Answer:")
        st.write(answer)