requirements.txt
streamlit
torch
numpy
pillow
gdown